- Added mock auth servise. - Extracted form to separate component. - Refactored.
