+++ package.json
@@ -15,8 +15,8 @@
   },
   "devDependencies": {
     "chai": "^4.1.2",
-    "jsdom": "^11.11.0",
     "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1"
+    "enzyme-adapter-react-16": "^1.1.1",
+    "jsdom": "^11.12.0"
   }
 }
