+++ src/app/application/Application.css
@@ -1,0 +1,8 @@
+.app {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(63, 81, 181, 0.07);
+    font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
+    font-weight: 400;
+}
