+++ src/app/login/Login.css
@@ -1,14 +1,5 @@
 @media only screen
 and (min-width: 0) {
-    .app {
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        background-color: rgba(63, 81, 181, 0.07);
-        font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-        font-weight: 400;
-    }
-
     .form-block {
         display: block;
         position: relative;
